Update workflow triggers and action versions in YAML files
Updated the event triggers for 'Code formatting' and 'Create GitHub Issues from TODOs, FIXMEs, etc.' workflows in `formatting.yml` and `todos-to-issues.yml` files respectively. The 'Code formatting' workflow will now be triggered on push events to the `main` branch for changes in `.cs`, `.js`, and `.css` files in the `src` directory. The 'Create GitHub Issues from TODOs, FIXMEs, etc.' workflow will be triggered on the completion of the 'Build and Test .NET projects' workflow for changes in `.cs` files in the `src` directory. Also, updated the version of the `actions/checkout` action used in the 'create_issues' job from `v3` to `v4`. Lastly, the 'TODO to Issue' step in the 'create_issues' job now uses specific identifiers for 'TODO' and 'FIXME' comments, with corresponding labels for the created issues.