Update formatting and todos-to-issues yml files (#8)
* Add GitHub Actions workflows and update .NET project settings

This commit introduces three new GitHub Actions workflows: `dotnet.yml` for automating the build process, `formatting.yml` for enforcing code formatting standards, and `todos-to-issues.yml` for creating GitHub issues from TODOs and FIXMEs. The solution file `BigBadBlog.sln` is updated to include these workflow files. Additionally, a `Global.json` file is added to specify the .NET SDK version and roll forward policy for the project.

* Updated workflows and solution file

Updated and added new workflows in `dotnet.yml`, `formatting.yml`, and `todos-to-issues.yml` for building .NET projects, code formatting, and creating GitHub issues from TODOs respectively. Also, updated the solution file `BigBadBlog.sln` to include these new workflow files in the "Solution Items" section.

* Update .NET setup path in dotnet.yml

Updated the path for `global-json-file` in the `Setup .NET` step of the `dotnet.yml` file. The new path is `./src/BigBadBlog/global.json`, indicating that the `global.json` file, which specifies the .NET SDK version, is now located in the `src/BigBadBlog` directory instead of the root directory.

* Update Paths in yml files

* Update .NET setup and solution file references

Updated the `dotnet.yml` file in the GitHub Actions workflow to change the `global-json-file` path in the "Setup .NET" step from `global.json` to `src/global.json`. Also, in the `BigBadBlog.sln` solution file, the reference to `Global.json` has been moved from the "Solution Items" project to the "0. Solution Files" project, indicating that the `Global.json` file is now considered part of the solution's root files.

* Updated case in Global.json reference

Updated the reference to the `Global.json` file in the "0. Solution Files" project section of the `BigBadBlog.sln` solution file to use lowercase 'g' in `global.json`. This change was made to address potential case sensitivity issues and adhere to naming conventions.

* Update path in Setup .NET step of dotnet.yml

Updated the path to the `global.json` file in the `Setup .NET` step of the `dotnet.yml` file from `src/global.json` to `./src/global.json`. This change corrects a potential path issue, ensuring the action correctly locates the `global.json` file in the `src` directory.

* Update workflow triggers and action versions in YAML files

Updated the event triggers for 'Code formatting' and 'Create GitHub Issues from TODOs, FIXMEs, etc.' workflows in `formatting.yml` and `todos-to-issues.yml` files respectively. The 'Code formatting' workflow will now be triggered on push events to the `main` branch for changes in `.cs`, `.js`, and `.css` files in the `src` directory. The 'Create GitHub Issues from TODOs, FIXMEs, etc.' workflow will be triggered on the completion of the 'Build and Test .NET projects' workflow for changes in `.cs` files in the `src` directory. Also, updated the version of the `actions/checkout` action used in the 'create_issues' job from `v3` to `v4`. Lastly, the 'TODO to Issue' step in the 'create_issues' job now uses specific identifiers for 'TODO' and 'FIXME' comments, with corresponding labels for the created issues.