Add `global.json` to specify .NET SDK version
Added a new file `global.json` to the "0. Solution Files" project in the `BigBadBlog.sln` solution. This file is used to specify the version of the .NET SDK to be used when building the project.

Updated SDK version and modified solution file paths

Updated the SDK version in `global.json` from `8.0.3` to `8.0.300`. Modified the `BigBadBlog.sln` file to update the paths for the GitHub workflows `dotnet.yml`, `formatting.yml`, and `todos-to-issues.yml`. Also added the `global.json` file to the solution items and replaced the project "0. Solution Files" with "Solution Items".

`Update GitHub Actions workflow for code formatting`

Updated the `formatting.yml` file to enhance the GitHub Actions workflow for code formatting. The workflow is now triggered on push events to the `main` branch and on `workflow_dispatch` events. It has been granted write permissions to the repository contents to enable changes to the repository files. The workflow is now concerned with `.cs`, `.js`, or `.css` files in the `src` directory. A new job named `code-formatting` has been defined.