Updated dependabot.yml and BigBadBlog.sln paths
Updated the `dependabot.yml` file to change the package ecosystem to "nuget", the directory for package manifests to "/", and the update schedule to "daily". In the `BigBadBlog.sln` file, updated the paths for the `.github\workflows\dotnet.yml`, `.github\workflows\formatting.yml`, and `.github\workflows\todos-to-issues.yml` files to be prefixed with "..\..\". Also added the `dependabot.yml` file to the solution items.